codeclimate: separate file
