+++ services/sidecar/src/simcore_service_sidecar/executor.py
@@ -26,51 +26,10 @@ from . import config, exceptions
 from .boot_mode import BootMode
 from .log_parser import LogType, monitor_logs_task
 from .rabbitmq import RabbitMQ
+from .task_shared_volume import TaskSharedVolumes
 from .utils import get_volume_mount_point
 
 log = logging.getLogger(__name__)
-
-
-@dataclass
-class TaskSharedVolumes:
-    input_folder: Path
-    output_folder: Path
-    log_folder: Path
-
-    @classmethod
-    def from_task(cls, task: aiopg.sa.result.RowProxy):
-        return cls(
-            config.SIDECAR_INPUT_FOLDER / f"{task.job_id}",
-            config.SIDECAR_OUTPUT_FOLDER / f"{task.job_id}",
-            config.SIDECAR_LOG_FOLDER / f"{task.job_id}",
-        )
-
-    def create(self) -> None:
-        for folder in [
-            self.input_folder,
-            self.output_folder,
-            self.log_folder,
-        ]:
-            if folder.exists():
-                shutil.rmtree(str(folder))
-            folder.mkdir(parents=True, exist_ok=True)
-
-    def delete(self) -> None:
-        for folder in [
-            self.input_folder,
-            self.output_folder,
-            self.log_folder,
-        ]:
-            if folder.exists():
-
-                def log_error(_, path, excinfo):
-                    log.warning(
-                        "Failed to remove %s [reason: %s]. Should consider pruning files in host later",
-                        path,
-                        excinfo,
-                    )
-
-                shutil.rmtree(str(folder), onerror=log_error)
 
 
 Version = Union[version.LegacyVersion, version.Version]
